+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## 0.11.0 (2018-11-27)
 * Don't fetch unused booking text
 * Show "choice - multiple" & "newsletter" fields as checkboxes
 * Get ready for "website" field